Removed nested column structure as parent component had no purpose and provided no benefit. Prettified files. Also removed cqc-registered-check.ts  as it was an unused class/file
