--- src/app/features/cqc-registered-question/cqc-registered-question-edit/cqc-registered-question-edit.component.html
@@ -1,134 +1,0 @@
-
-<form novalidate
-      (ngSubmit)="onSubmit()"
-      [formGroup]="cqcRegisteredQuestionForm">
-
-  <div class="form-group mb-5">
-    <div class="col-12 p-0">
-      <label class="main-body-text-bold mb-3 d-block">Are you regulated by the Care Quality Commission (CQC)?</label>
-
-      <div class="row mb-5">
-        <div class="col-12">
-          <div class="multiple-choice">
-            <input type="radio" id="reqQuestionYes" value="true" formControlName="registeredQuestionSelected" />
-            <label class="main-body-text p-0" for="reqQuestionYes">Yes</label>
-          </div>
-        </div>
-
-        <div *ngIf="registeredQuestionSelectedValue === 'true'" class="panel panel-border-narrow col-10 col-lg-6"
-             [ngClass]="{'is-invalid': (!isSubmitted && (cqcRegisteredPostcode.errors || cqcRegisteredLocationId.errors || cqcRegisteredGroup.errors)) ||
-                                      (isSubmitted && (cqcRegisteredPostcode.errors || cqcRegisteredLocationId.errors || cqcRegisteredGroup.errors)) ||
-                                      cqcRegisteredGroup.hasError('bothHaveContent') ||
-                                      (this.cqcPostcodeApiError || this.cqclocationApiError) }">
-
-          <div formGroupName="cqcRegisteredGroup">
-            <div class="form-group">
-              <label class="form-label" for="postcodeYes">Enter your postcode</label>
-
-              <span class="invalid-feedback">
-                <span *ngIf="!submittedCqcRegPostcode" class="main-body-text mb-2">{{ cqcRegPostcodeMessage }}</span>
-                <!-- <span *ngIf="isSubmitted && !submittedCqcRegPostcode && !submittedCqcRegLocationId" class="main-body-text mb-2">
-                    Please enter a postcode or Location ID.
-                </span> -->
-                <span *ngIf="submittedCqcRegPostcode && cqcRegisteredPostcode.hasError('maxlength')" class="main-body-text mb-2">
-                    Your postcode must be no longer than 8 characters.
-                </span>
-                <span *ngIf="submittedCqcRegPostcode && cqcRegisteredGroup.hasError('bothHaveContent')" class="main-body-text mb-2">
-                  Both have content
-                </span>
-                <span *ngIf="this.cqcPostcodeApiError" class="main-body-text mb-2">
-                  {{ this.cqcPostcodeApiError }}
-                </span>
-              </span>
-
-              <div class="form-control-feedback">
-
-                <input placeholder="postcode" type="text" id="postcodeYes" class="form-control col-12" formControlName="cqcRegisteredPostcode" />
-
-
-              </div>
-            </div>
-
-            <div class="form-group mt-2">
-              <p class="label mt-4">or</p>
-
-              <label class="form-label" for="locationId">Enter your location ID number</label>
-
-              <span class="invalid-feedback">
-                <span *ngIf="!submittedCqcRegLocationId" class="main-body-text mb-2">{{ cqcRegLocationIdMessage }}</span>
-                <!-- <span *ngIf="isSubmitted && !submittedCqcRegPostcode && !submittedCqcRegLocationId" class="main-body-text mb-2">
-                    Please enter a postcode or Location ID.
-                </span> -->
-                  <span *ngIf="submittedCqcRegLocationId && cqcRegisteredLocationId.hasError('maxlength')" class="main-body-text mb-2">
-                      Your Location ID must be no longer than 50 characters.
-                  </span>
-                  <span *ngIf="submittedCqcRegPostcode && cqcRegisteredGroup.hasError('bothHaveContent')" class="main-body-text mb-2">
-                    Both have content
-                  </span>
-                  <span *ngIf="this.cqclocationApiError" class="main-body-text mb-2">
-                    {{ this.cqclocationApiError }}
-                  </span>
-              </span>
-
-              <input placeholder="location" type="text" id="location" class="form-control col-12" formControlName="locationId" />
-
-            </div>
-
-          </div>
-
-        </div>
-      </div>
-
-      <div class="row">
-        <div class="col-12">
-          <div class="multiple-choice">
-            <input type="radio" id="regQuestionNo" value="false" formControlName="registeredQuestionSelected" />
-
-            <label class="main-body-text p-0" for="regQuestionNo">No</label>
-          </div>
-        </div>
-
-        <div *ngIf="registeredQuestionSelectedValue === 'false'" class="panel panel-border-narrow col-10 col-lg-6"
-             [ngClass]="{'is-invalid': nonCqcRegPostcodeMessage || this.nonCqcPostcodeApiError }">
-          <div class="field-group">
-
-            <label class="form-label" for="postcodeNo">What is your postcode?</label>
-
-            <span class="invalid-feedback">
-              <span *ngIf="nonCqcRegPostcodeMessage" class="main-body-text mb-2">{{ nonCqcRegPostcodeMessage }}</span>
-              <span *ngIf="submittedCqcRegPostcode && notRegisteredPostcode.hasError('maxlength')" class="main-body-text mb-2">
-                The Postcode must be no longer than 8 characters.
-              </span>
-              <span *ngIf="this.nonCqcPostcodeApiError" class="main-body-text mb-2">
-                {{ this.nonCqcPostcodeApiError }}
-              </span>
-            </span>
-
-            <input placeholder="postcode" type="text" id="postcodeNo" class="form-control col-12" formControlName="notRegisteredPostcode"  />
-
-          </div>
-        </div>
-
-      </div>
-    </div>
-  </div>
-
-  <div class="btn-container">
-    <button class="btn btn-primary" type="submit">Continue</button>
-  </div>
-
-</form>
-
-<!-- <br>Api Error: {{ this.apiError }}
-<br>Dirty: {{ this.cqcRegisteredQuestionForm.dirty }}
-<br>Touched: {{ this.cqcRegisteredQuestionForm.touched }}
-<br>Valid: {{ this.cqcRegisteredQuestionForm.valid }}
-<br>Value: {{ this.cqcRegisteredQuestionForm.value | json }}
-<br>isSubmitted {{ this.isSubmitted }}
-<br>submittedCqcRegPostcode: {{ this.submittedCqcRegPostcode }}
-<br>submittedCqcRegLocationId: {{ this.submittedCqcRegLocationId }}
-<br>cqcRegPostcodeMessage: {{ this.cqcRegPostcodeMessage }}
-<br >cqcRegLocationIdMessage: {{ this.cqcRegLocationIdMessage }}
-<br>postcode errors: {{ cqcRegisteredPostcode.errors | json }}
-<br>location ID errors: {{ cqcRegisteredLocationId.errors | json }} -->
-
