--- src/app/features/cqc-registered-question/cqc-registered-question-edit/cqc-regsitered-check.ts
@@ -1,9 +1,0 @@
-export class CqcRegisteredQuestionEnteredLocation {
-
-  constructor(
-    public registeredQuestionSelected = '',
-    public cqcRegisteredPostcode = '',
-    public locationId = '',
-    public notRegisteredPostcode = ''
-  ) {}
-}
